Rename files by template
