+++ testing/tests/DevExpress.ui.widgets/accordion.genericTheme.tests.js
@@ -1,5 +1,5 @@
 import 'common.css!';
 import 'generic_light.css!';
 
-import { runThemesSharedTests } from './accordionParts/accordion.sharedThemes.tests.js';
+import { runThemesSharedTests } from './accordionParts/accordion.themes.sharedTests.js';
 runThemesSharedTests('generic_light');
