+++ testing/tests/DevExpress.ui.widgets/accordion.materialTheme.tests.js
@@ -1,5 +1,5 @@
 import 'common.css!';
 import 'material_blue_light.css!';
 
-import { runThemesSharedTests } from './accordionParts/accordion.sharedThemes.tests.js';
+import { runThemesSharedTests } from './accordionParts/accordion.themes.sharedTests.js';
 runThemesSharedTests('material_blue_light');
