+++ testing/tests/DevExpress.ui.widgets/accordionParts/accordion.themes.sharedTests.js
